+++ pallets/subnet_emission/Cargo.toml
@@ -3,7 +3,7 @@ name = "pallet-subnet-emission"
 version = "1.0.1"
 description = "FRAME pallet for runtime logic of Subspace Blockchain."
 authors = ["Commune Community"]
-homepage = "https://commune.com"
+homepage = "https://communeai.org"
 edition = "2021"
 license = "Unlicense"
 publish = false
