+++ pallets/subnet_emission/src/consensus/treasury.rs
@@ -2,7 +2,7 @@ use crate::EmissionError;
 use core::marker::PhantomData;
 use pallet_subspace::{Config, Pallet as PalletSubspace};
 
-use crate::subnet_consensus::util::params::AccountKey;
+use crate::consensus::util::params::AccountKey;
 // Code structure to reflect other consensus types, this code is ready for additional features.
 // Whenever needed.
 pub struct TreasuryEpoch<T: Config> {
